requirements.txt
aiohttp==1.0.5
async-timeout==1.4.0
cffi==1.11.1
chardet==3.0.4
discord.py==0.16.12
future==0.16.0
multidict==3.2.0
pycparser==2.18
PyNaCl==1.0.1
PyYAML==3.12
six==1.11.0
websockets==3.4
youtube-dl==2017.10.15.1
